requirements.txt
anyconfig==0.13.0
apache-sedona==1.4.1
astropy==5.1
beautifulsoup4==4.11.1
fastapi[all]==0.100.0
numpy==1.23.4
pandas[parquet, feather]==2.0.3
PyYAML==6.0
requests==2.28.1
tqdm==4.65.0
wget==3.2
